PrjctMgmtOffc_Project/src/main/java/com/fse/pmo/dao/ProjectDaoImpl.java
package com.fse.pmo.dao;

import java.util.ArrayList;
import java.util.List;

import javax.persistence.EntityManagerFactory;
import javax.persistence.criteria.CriteriaBuilder;
import javax.persistence.criteria.CriteriaQuery;
import javax.persistence.criteria.Root;

import org.hibernate.Session;
import org.hibernate.SessionFactory;
import org.hibernate.Transaction;
import org.hibernate.query.Query;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Repository;


import com.fse.pmo.model.PmoProject;

@Repository
public class ProjectDaoImpl implements ProjectDao {

	@Autowired 
	private EntityManagerFactory entityManagerFactory;
	
	@Override
	public Integer saveUpadteProject(PmoProject project) {
		Integer result = 0;
		SessionFactory sessionFactory = entityManagerFactory.unwrap(SessionFactory.class);
		Session session = null;
		try {
			session = sessionFactory.openSession();
			Transaction beginTransaction = session.beginTransaction();
			session.saveOrUpdate(project);
			result = 1; 
			beginTransaction.commit();
		} catch (Exception e) {
			e.printStackTrace();
		}finally {
			session.close();
		}
		
		return result;
		
	}

	@Override
	public List<PmoProject> getAllProjects() {
		List<PmoProject> results = new ArrayList<>();
		SessionFactory sessionFactory = entityManagerFactory.unwrap(SessionFactory.class);
		Session session = null;
		try {
			session=sessionFactory.openSession();
			CriteriaBuilder cb = session.getCriteriaBuilder();
			CriteriaQuery<PmoProject> createQuery = cb.createQuery(PmoProject.class);
			Root<PmoProject> root = createQuery.from(PmoProject.class);
			createQuery.orderBy(cb.desc(root.get("project_id")));
			createQuery.select(root);
			 
			Query<PmoProject> query = session.createQuery(createQuery);
			results = query.getResultList();
		} catch (Exception e) {
			e.printStackTrace();
		}finally {
			session.close();
		}
		return results;
	}

	@Override
	public PmoProject getProjectById(Integer projectId) {
		PmoProject project = null;
		SessionFactory sessionFactory = entityManagerFactory.unwrap(SessionFactory.class);
		Session session = null;
		try {
			session = sessionFactory.openSession();
			Transaction beginTransaction = session.beginTransaction();
			project = (PmoProject)session.get(PmoProject.class, projectId);
			beginTransaction.commit();
		} catch (Exception e) {
			e.printStackTrace();
		}finally {
			session.close();
		}
		return project;
	}

	@Override
	public Integer deleteProjectById(Integer projectId) {
		SessionFactory sessionFactory = entityManagerFactory.unwrap(SessionFactory.class);
		Session session=null;
		try {
			session=sessionFactory.openSession();
			Transaction beginTransaction = session.beginTransaction();
			PmoProject project = (PmoProject)session.get(PmoProject.class, projectId);
			//Remove user reference
			//PmoUser updateUser = project.getUser();
			//updateUser.setProject(null);
			//session.saveOrUpdate(updateUser);
			session.delete(project);
			beginTransaction.commit();
			
		} catch (Exception e) {
			e.printStackTrace();
		}finally {
			session.close();
		}
		return null;
	}
	

}
